% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_quarto-r6.R
\name{check_deploy_quarto_docx}
\alias{check_deploy_quarto_docx}
\title{R6 classes}
\description{
R6 classes

R6 classes
}
\details{
Check classes
}
\section{Super classes}{
\code{jrHealthCheckConnect::logger} -> \code{jrHealthCheckConnect::base_check} -> \code{check_deploy_quarto_docx}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-check_deploy_quarto_docx-check}{\code{check_deploy_quarto_docx$check()}}
\item \href{#method-check_deploy_quarto_docx-clone}{\code{check_deploy_quarto_docx$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="logger" data-id="get_log"><a href='../../jrHealthCheckConnect/html/logger.html#method-logger-get_log'><code>jrHealthCheckConnect::logger$get_log()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="logger" data-id="start_logger"><a href='../../jrHealthCheckConnect/html/logger.html#method-logger-start_logger'><code>jrHealthCheckConnect::logger$start_logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="logger" data-id="stop_logger"><a href='../../jrHealthCheckConnect/html/logger.html#method-logger-stop_logger'><code>jrHealthCheckConnect::logger$stop_logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="base_check" data-id="info"><a href='../../jrHealthCheckConnect/html/base_check.html#method-base_check-info'><code>jrHealthCheckConnect::base_check$info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="base_check" data-id="initialize"><a href='../../jrHealthCheckConnect/html/base_check.html#method-base_check-initialize'><code>jrHealthCheckConnect::base_check$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_deploy_quarto_docx-check"></a>}}
\if{latex}{\out{\hypertarget{method-check_deploy_quarto_docx-check}{}}}
\subsection{Method \code{check()}}{
Checks deployment of Quarto document with Docx output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{check_deploy_quarto_docx$check(account = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account}}{Connect username}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_deploy_quarto_docx-clone"></a>}}
\if{latex}{\out{\hypertarget{method-check_deploy_quarto_docx-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{check_deploy_quarto_docx$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
