% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_quarto-r6.R
\name{check_deploy_quarto_observable}
\alias{check_deploy_quarto_observable}
\title{R6 classes}
\description{
R6 classes

R6 classes
}
\details{
Check classes
}
\section{Super classes}{
\code{\link[uatBase:logger]{uatBase::logger}} -> \code{\link[uatBase:base_check]{uatBase::base_check}} -> \code{check_deploy_quarto_observable}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-check_deploy_quarto_observable-check}{\code{check_deploy_quarto_observable$check()}}
\item \href{#method-check_deploy_quarto_observable-clone}{\code{check_deploy_quarto_observable$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="uatBase" data-topic="logger" data-id="get_log"><a href='../../uatBase/html/logger.html#method-logger-get_log'><code>uatBase::logger$get_log()</code></a></span></li>
<li><span class="pkg-link" data-pkg="uatBase" data-topic="logger" data-id="start_logger"><a href='../../uatBase/html/logger.html#method-logger-start_logger'><code>uatBase::logger$start_logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="uatBase" data-topic="logger" data-id="stop_logger"><a href='../../uatBase/html/logger.html#method-logger-stop_logger'><code>uatBase::logger$stop_logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="uatBase" data-topic="base_check" data-id="info"><a href='../../uatBase/html/base_check.html#method-base_check-info'><code>uatBase::base_check$info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="uatBase" data-topic="base_check" data-id="initialize"><a href='../../uatBase/html/base_check.html#method-base_check-initialize'><code>uatBase::base_check$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_deploy_quarto_observable-check"></a>}}
\if{latex}{\out{\hypertarget{method-check_deploy_quarto_observable-check}{}}}
\subsection{Method \code{check()}}{
Checks deployment of Quarto document with PDF output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{check_deploy_quarto_observable$check(debug_level, account = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{debug_level}}{See check() for details}

\item{\code{account}}{Connect username}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_deploy_quarto_observable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-check_deploy_quarto_observable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{check_deploy_quarto_observable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
