% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_rmd-r6.R
\name{check_deploy_rmd_pdf}
\alias{check_deploy_rmd_pdf}
\alias{check_deploy_rmd_html}
\alias{check_deploy_rmd_doc}
\title{R6 classes}
\description{
R6 classes

R6 classes
}
\details{
Check classes
}
\section{Super classes}{
\code{jrHealthCheckConnect::logger} -> \code{jrHealthCheckConnect::base_check} -> \code{check_deploy_rmd_pdf}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-check_deploy_rmd_pdf-check}{\code{check_deploy_rmd_pdf$check()}}
\item \href{#method-check_deploy_rmd_pdf-clone}{\code{check_deploy_rmd_pdf$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="logger" data-id="get_log"><a href='../../jrHealthCheckConnect/html/logger.html#method-logger-get_log'><code>jrHealthCheckConnect::logger$get_log()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="logger" data-id="start_logger"><a href='../../jrHealthCheckConnect/html/logger.html#method-logger-start_logger'><code>jrHealthCheckConnect::logger$start_logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="logger" data-id="stop_logger"><a href='../../jrHealthCheckConnect/html/logger.html#method-logger-stop_logger'><code>jrHealthCheckConnect::logger$stop_logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="base_check" data-id="info"><a href='../../jrHealthCheckConnect/html/base_check.html#method-base_check-info'><code>jrHealthCheckConnect::base_check$info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="base_check" data-id="initialize"><a href='../../jrHealthCheckConnect/html/base_check.html#method-base_check-initialize'><code>jrHealthCheckConnect::base_check$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_deploy_rmd_pdf-check"></a>}}
\if{latex}{\out{\hypertarget{method-check_deploy_rmd_pdf-check}{}}}
\subsection{Method \code{check()}}{
Checks deployment of an R Markdown document with PDF output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{check_deploy_rmd_pdf$check()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_deploy_rmd_pdf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-check_deploy_rmd_pdf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{check_deploy_rmd_pdf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{jrHealthCheckConnect::logger} -> \code{jrHealthCheckConnect::base_check} -> \code{check_deploy_rmd_html}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-check_deploy_rmd_html-check}{\code{check_deploy_rmd_html$check()}}
\item \href{#method-check_deploy_rmd_html-clone}{\code{check_deploy_rmd_html$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="logger" data-id="get_log"><a href='../../jrHealthCheckConnect/html/logger.html#method-logger-get_log'><code>jrHealthCheckConnect::logger$get_log()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="logger" data-id="start_logger"><a href='../../jrHealthCheckConnect/html/logger.html#method-logger-start_logger'><code>jrHealthCheckConnect::logger$start_logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="logger" data-id="stop_logger"><a href='../../jrHealthCheckConnect/html/logger.html#method-logger-stop_logger'><code>jrHealthCheckConnect::logger$stop_logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="base_check" data-id="info"><a href='../../jrHealthCheckConnect/html/base_check.html#method-base_check-info'><code>jrHealthCheckConnect::base_check$info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="base_check" data-id="initialize"><a href='../../jrHealthCheckConnect/html/base_check.html#method-base_check-initialize'><code>jrHealthCheckConnect::base_check$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_deploy_rmd_html-check"></a>}}
\if{latex}{\out{\hypertarget{method-check_deploy_rmd_html-check}{}}}
\subsection{Method \code{check()}}{
Checks deployment of an R Markdown document with HTML output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{check_deploy_rmd_html$check()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_deploy_rmd_html-clone"></a>}}
\if{latex}{\out{\hypertarget{method-check_deploy_rmd_html-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{check_deploy_rmd_html$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{jrHealthCheckConnect::logger} -> \code{jrHealthCheckConnect::base_check} -> \code{check_deploy_rmd_word}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-check_deploy_rmd_word-check}{\code{check_deploy_rmd_doc$check()}}
\item \href{#method-check_deploy_rmd_word-clone}{\code{check_deploy_rmd_doc$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="logger" data-id="get_log"><a href='../../jrHealthCheckConnect/html/logger.html#method-logger-get_log'><code>jrHealthCheckConnect::logger$get_log()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="logger" data-id="start_logger"><a href='../../jrHealthCheckConnect/html/logger.html#method-logger-start_logger'><code>jrHealthCheckConnect::logger$start_logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="logger" data-id="stop_logger"><a href='../../jrHealthCheckConnect/html/logger.html#method-logger-stop_logger'><code>jrHealthCheckConnect::logger$stop_logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="base_check" data-id="info"><a href='../../jrHealthCheckConnect/html/base_check.html#method-base_check-info'><code>jrHealthCheckConnect::base_check$info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jrHealthCheckConnect" data-topic="base_check" data-id="initialize"><a href='../../jrHealthCheckConnect/html/base_check.html#method-base_check-initialize'><code>jrHealthCheckConnect::base_check$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_deploy_rmd_word-check"></a>}}
\if{latex}{\out{\hypertarget{method-check_deploy_rmd_word-check}{}}}
\subsection{Method \code{check()}}{
Checks deployment of an R Markdown document with Word Docx output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{check_deploy_rmd_doc$check()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_deploy_rmd_word-clone"></a>}}
\if{latex}{\out{\hypertarget{method-check_deploy_rmd_word-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{check_deploy_rmd_doc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
