% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_config}
\alias{create_config}
\title{Create test config}
\usage{
create_config(dir = ".", file = "config-uat.yml")
}
\arguments{
\item{dir}{Directory of the config file}

\item{file}{Config file name}
}
\description{
Some PSC servers intentionally won't have some features implemented.
Tests can be skipped using a config-uat.yml file.
}
\details{
This function creates an example config. By default all tests are TRUE.

If a test is missing from the config file, it is assume to be TRUE.
Therefore, the config file can be quite short and just list exceptions.
If the config file is missing, then all tests are carried out.
}
