% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_python.R
\name{deploy_python}
\alias{deploy_python}
\title{Deploy Python output}
\usage{
deploy_python(
  python_dir,
  python_files = c("app.py", "index.qmd"),
  rsconnect_type = c("api", "streamlit", "quarto"),
  debug_level = debug_level
)
}
\arguments{
\item{python_dir}{Location of python deployment files}

\item{python_files}{Either an app or a qmd file}

\item{rsconnect_type}{Type of deployment}

\item{debug_level}{Integer, 0 to 2.}
}
\description{
Deploy Python output
}
\details{
Python deploy function
All python content must have a requirements.txt file and either a app.py or index.qmd.
\code{index.qmd} is for quarto only. General idea is to
\itemize{
\item copy dir of contents to R_TMP_DIR
\item Use the python interface to deploy
\item Use the python interface to grab the guid
\item Clean up.
@noRd
}
}
