% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-helpers.R
\name{get_quarto_old_users}
\alias{get_quarto_old_users}
\alias{get_quarto_user_roles}
\title{Quarto Helper Functions}
\usage{
get_quarto_old_users(out)

get_quarto_user_roles(out)
}
\arguments{
\item{out}{Output from check}
}
\description{
In general not used by users.
}
