% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_users.R
\name{sanitize}
\alias{sanitize}
\title{Santitize {audit.connect} object}
\usage{
sanitize(audit_connect_object_arg)
}
\arguments{
\item{audit_connect_object_arg}{An object from audit.connect::check()}
}
\description{
This function wipes any user-identifying data from an {audit.connect} object,
but inserts a column which contains the URL for each user in the Posit Connect admin panel
}
