% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Run UAT Checks}
\usage{
check(dir = ".", file = "config-uat.yml")
}
\arguments{
\item{dir}{directory location of the the config file}

\item{file}{config file name}
}
\description{
This functions runs all UAT tests.
To skip tests, set check to `no` in the config yaml file.
See `create_config()`
}
