% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Run Posit Checks}
\usage{
check(
  server = NULL,
  token = NULL,
  dir = ".",
  debug_level = 0:2,
  suppress_users = FALSE
)
}
\arguments{
\item{server}{Connect server. If NULL, use the ENV variable CONNECT_SERVER}

\item{token}{Connect api token. If NULL, use the ENV variable CONNECT_API_KEY}

\item{dir}{directory location of the the config file}

\item{debug_level}{Integer, 0 to 2.}

\item{suppress_users}{Whether to suppress output of Connect users' data.}
}
\description{
This functions runs all Posit tests.
To skip tests, set check to \code{no} in the config yaml file.
See \code{create_config()}
}
\details{
Debug level description
\itemize{
\item 0: clean-up all files; suppress all noise
\item 1: clean-up all files, but display build steps
\item 2: No clean-up (on connect and on disk) and display build steps
}
}
